requirements.txt
streamlit>=1.44.0
google-cloud-bigquery>=3.30.0
google-cloud-storage>=2.18.0
google-cloud-aiplatform>=1.71.0
pydantic>=2.10.0
python-dotenv>=1.0.1
fastapi>=0.116.0
uvicorn>=0.35.0
python-multipart>=0.0.20
